+++ test_runner/src/main/kotlin/ftl/gc/GcStorage.kt
@@ -5,9 +5,13 @@ import com.google.cloud.storage.Storage
 import com.google.cloud.storage.StorageOptions
 import com.google.cloud.storage.contrib.nio.testing.LocalStorageHelper
 import ftl.args.AndroidArgs
+import ftl.args.IArgs
 import ftl.args.IosArgs
 import ftl.config.FtlConstants
 import ftl.config.FtlConstants.GCS_PREFIX
+import ftl.reports.xml.model.JUnitTestResult
+import ftl.reports.xml.parseIosXml
+import ftl.reports.xml.xmlToString
 import ftl.util.Utils.fatalError
 import ftl.util.Utils.join
 import java.io.File
@@ -40,6 +44,23 @@ object GcStorage {
             runGcsPath = runGcsPath
         )
 
+    fun uploadJunitXml(testResult: JUnitTestResult, args: IArgs) {
+        if (args.smartFlankGcsPath.isEmpty()) return
+
+        // bucket/path/to/object
+        val rawPath = args.smartFlankGcsPath.drop(GCS_PREFIX.length)
+        val bucket = rawPath.substringBefore('/')
+        val name = rawPath.substringAfter('/')
+
+        val fileBlob = BlobInfo.newBuilder(bucket, name).build()
+
+        try {
+            storage.create(fileBlob, testResult.xmlToString().toByteArray())
+        } catch (e: Exception) {
+            fatalError(e)
+        }
+    }
+
     fun uploadAppApk(args: AndroidArgs, gcsBucket: String, runGcsPath: String): String =
         upload(args.appApk, gcsBucket, runGcsPath)
 
@@ -60,6 +81,16 @@ object GcStorage {
     fun downloadTestApk(args: AndroidArgs): String =
         download(args.testApk)
 
+    // junit xml may not exist. ignore error if it doesn't exist
+    fun downloadJunitXml(args: IArgs): JUnitTestResult? {
+        val oldXmlPath = download(args.smartFlankGcsPath, ignoreError = true)
+        if (oldXmlPath.isNotEmpty()) {
+            return parseIosXml(Paths.get(oldXmlPath))
+        }
+
+        return null
+    }
+
     private fun upload(file: String, fileBytes: ByteArray, rootGcsBucket: String, runGcsPath: String): String {
         val fileName = Paths.get(file).fileName.toString()
         val gcsFilePath = GCS_PREFIX + join(rootGcsBucket, runGcsPath, fileName)
@@ -76,12 +107,12 @@ object GcStorage {
         return gcsFilePath
     }
 
-    private fun download(gcsUriString: String): String {
+    private fun download(gcsUriString: String, ignoreError: Boolean = false): String {
         val gcsURI = URI.create(gcsUriString)
         val bucket = gcsURI.authority
         val path = gcsURI.path.drop(1) // Drop leading slash
 
-        val outputFile = File.createTempFile("apk", null)
+        val outputFile = File.createTempFile("tmp", null)
         outputFile.deleteOnExit()
 
         try {
@@ -91,6 +122,7 @@ object GcStorage {
             output.channel.transferFrom(readChannel, 0, Long.MAX_VALUE)
             output.close()
         } catch (e: Exception) {
+            if (ignoreError) return ""
             fatalError(e)
         }