+++ test_runner/src/test/kotlin/ftl/args/AndroidArgsTest.kt
@@ -202,6 +202,7 @@ AndroidArgs
     flank:
       testShards: 7
       repeatTests: 8
+      smartFlankGcsPath:${' '}
       test-targets-always-run:
         - class example.Test#grantPermission
         - class example.Test#grantPermission2