+++ test_runner/src/test/kotlin/ftl/run/GenericTestRunnerTest.kt
@@ -23,7 +23,7 @@ class GenericTestRunnerTest {
     @Test
     fun testBeforeRunMessage1() {
         val result = beforeRunMessage(createMock(1, listOf(listOf(""))))
-        assert(result, "  1 test / 1 shard = 1 test per shard\n")
+        assert(result, "  1 test / 1 shard\n")
     }
 
     @Test
@@ -31,7 +31,7 @@ class GenericTestRunnerTest {
         val result = beforeRunMessage(createMock(2, listOf(listOf(""))))
         assert(
             result, """
-  1 test / 1 shard = 1 test per shard
+  1 test / 1 shard
   Running 2x
     2 total shards
     2 total tests
@@ -49,7 +49,7 @@ class GenericTestRunnerTest {
         )
         assert(
             result, """
-  6 tests / 6 shards = 1 test per shard
+  6 tests / 6 shards
   Running 2x
     12 total shards
     12 total tests
@@ -62,7 +62,7 @@ class GenericTestRunnerTest {
         val result = beforeRunMessage(createMock(100, listOf(listOf("", "", "", "", ""), listOf("", "", "", "", ""))))
         assert(
             result, """
-  10 tests / 2 shards = 5 tests per shard
+  10 tests / 2 shards
   Running 100x
     200 total shards
     1000 total tests